pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "gibbsq"
version = "0.1.0"
description = "Stability verification of softmax-routed queueing networks"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.24",
    "scipy>=1.11",
    "matplotlib>=3.8",
    "hydra-core>=1.3",
    "omegaconf>=2.3",
    "pandas>=2.0",
    "pyarrow>=14.0",
    "jax>=0.4.20",
    "jaxlib>=0.4.20",
    "wandb>=0.16.0",
    "optax>=0.1.7",
    "equinox>=0.11.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "numba>=0.59",
    "hypothesis>=6.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
